requirements.txt
firecrawl-py
pgvector
psycopg2-binary
sqlalchemy
phidata
anthropic
python-dotenv
streamlit
openai
